// src/app/data-access/api/api-service-base.service.ts
export class FilterDescriptor implements IFilterDescriptor {
  logic?: string | undefined;
  filters?: FilterDescriptor[] | undefined;
  field?: string | undefined;
  operator?: string | undefined;
  value?: any | undefined;

  constructor(data?: IFilterDescriptor) {
    if (data) {
      for (var property in data) {
        if (data.hasOwnProperty(property))
          (this as any)[property] = (data as any)[property];
      }
    }
  }

  init(_data?: any) {
    if (_data) {
      this.logic = _data['logic'];
      if (Array.isArray(_data['filters'])) {
        this.filters = [] as any;
        for (let item of _data['filters'])
          this.filters!.push(FilterDescriptor.fromJS(item));
      }
      this.field = _data['field'];
      this.operator = _data['operator'];
      this.value = _data['value'];
    }
  }

  static fromJS(data: any): FilterDescriptor {
    data = typeof data === 'object' ? data : {};
    let result = new FilterDescriptor();
    result.init(data);
    return result;
  }

  toJSON(data?: any) {
    data = typeof data === 'object' ? data : {};
    data['logic'] = this.logic;
    if (Array.isArray(this.filters)) {
      data['filters'] = [];
      for (let item of this.filters)
        data['filters'].push(item ? item.toJSON() : (undefined as any));
    }
    data['field'] = this.field;
    data['operator'] = this.operator;
    data['value'] = this.value;
    return data;
  }
}

export interface IFilterDescriptor {
  logic?: string | undefined;
  filters?: FilterDescriptor[] | undefined;
  field?: string | undefined;
  operator?: string | undefined;
  value?: any | undefined;
}

export class GetProductsQuery implements IGetProductsQuery {
  shopId?: string | undefined;
  search?: string | undefined;
  filter?: FilterDescriptor;
  page?: number;
  pageSize?: number;

  constructor(data?: IGetProductsQuery) {
    if (data) {
      for (var property in data) {
        if (data.hasOwnProperty(property))
          (this as any)[property] = (data as any)[property];
      }
    }
  }

  init(_data?: any) {
    if (_data) {
      this.shopId = _data['shopId'];
      this.search = _data['search'];
      this.filter = _data['filter']
        ? FilterDescriptor.fromJS(_data['filter'])
        : (undefined as any);
      this.page = _data['page'];
      this.pageSize = _data['pageSize'];
    }
  }

  static fromJS(data: any): GetProductsQuery {
    data = typeof data === 'object' ? data : {};
    let result = new GetProductsQuery();
    result.init(data);
    return result;
  }

  toJSON(data?: any) {
    data = typeof data === 'object' ? data : {};
    data['shopId'] = this.shopId;
    data['search'] = this.search;
    data['filter'] = this.filter ? this.filter.toJSON() : (undefined as any);
    data['page'] = this.page;
    data['pageSize'] = this.pageSize;
    return data;
  }
}

export interface IGetProductsQuery {
  shopId?: string | undefined;
  search?: string | undefined;
  filter?: FilterDescriptor;
  page?: number;
  pageSize?: number;
}
// ...
export class ProductDto implements IProductDto {
  id?: string;
  name?: string | undefined;
  sku?: string | undefined;
  slug?: string | undefined;
  description?: string | undefined;
  summary?: string | undefined;
  status?: string | undefined;
  viewCount?: number | undefined;
  brandId?: string | undefined;
  categoryId?: string | undefined;
  shopId?: string | undefined;
  baseAttributes?: string | undefined;
  createdAt?: Date | undefined;
  updatedAt?: Date | undefined;
  images?: ProductImageDto[] | undefined;

  constructor(data?: IProductDto) {
    if (data) {
      for (var property in data) {
        if (data.hasOwnProperty(property))
          (this as any)[property] = (data as any)[property];
      }
    }
  }

  init(_data?: any) {
    if (_data) {
      this.id = _data['id'];
      this.name = _data['name'];
      this.sku = _data['sku'];
      this.slug = _data['slug'];
      this.description = _data['description'];
      this.summary = _data['summary'];
      this.status = _data['status'];
      this.viewCount = _data['viewCount'];
      this.brandId = _data['brandId'];
      this.categoryId = _data['categoryId'];
      this.shopId = _data['shopId'];
      this.baseAttributes = _data['baseAttributes'];
      this.createdAt = _data['createdAt']
        ? new Date(_data['createdAt'].toString())
        : (undefined as any);
      this.updatedAt = _data['updatedAt']
        ? new Date(_data['updatedAt'].toString())
        : (undefined as any);
      if (Array.isArray(_data['images'])) {
        this.images = [] as any;
        for (let item of _data['images'])
          this.images!.push(ProductImageDto.fromJS(item));
      }
    }
  }

  static fromJS(data: any): ProductDto {
    data = typeof data === 'object' ? data : {};
    let result = new ProductDto();
    result.init(data);
    return result;
  }

  toJSON(data?: any) {
    data = typeof data === 'object' ? data : {};
    data['id'] = this.id;
    data['name'] = this.name;
    data['sku'] = this.sku;
    data['slug'] = this.slug;
    data['description'] = this.description;
    data['summary'] = this.summary;
    data['status'] = this.status;
    data['viewCount'] = this.viewCount;
    data['brandId'] = this.brandId;
    data['categoryId'] = this.categoryId;
    data['shopId'] = this.shopId;
    data['baseAttributes'] = this.baseAttributes;
    data['createdAt'] = this.createdAt
      ? this.createdAt.toISOString()
      : (undefined as any);
    data['updatedAt'] = this.updatedAt
      ? this.updatedAt.toISOString()
      : (undefined as any);
    if (Array.isArray(this.images)) {
      data['images'] = [];
      for (let item of this.images)
        data['images'].push(item ? item.toJSON() : (undefined as any));
    }
    return data;
  }
}

export interface IProductDto {
  id?: string;
  name?: string | undefined;
  sku?: string | undefined;
  slug?: string | undefined;
  description?: string | undefined;
  summary?: string | undefined;
  status?: string | undefined;
  viewCount?: number | undefined;
  brandId?: string | undefined;
  categoryId?: string | undefined;
  shopId?: string | undefined;
  baseAttributes?: string | undefined;
  createdAt?: Date | undefined;
  updatedAt?: Date | undefined;
  images?: ProductImageDto[] | undefined;
}
// ...
export class ProductImageDto implements IProductImageDto {
  id?: string;
  productId?: string;
  imageUrl?: string | undefined;
  displayOrder?: number;
  isMain?: boolean;

  constructor(data?: IProductImageDto) {
    if (data) {
      for (var property in data) {
        if (data.hasOwnProperty(property))
          (this as any)[property] = (data as any)[property];
      }
    }
  }

  init(_data?: any) {
    if (_data) {
      this.id = _data['id'];
      this.productId = _data['productId'];
      this.imageUrl = _data['imageUrl'];
      this.displayOrder = _data['displayOrder'];
      this.isMain = _data['isMain'];
    }
  }

  static fromJS(data: any): ProductImageDto {
    data = typeof data === 'object' ? data : {};
    let result = new ProductImageDto();
    result.init(data);
    return result;
  }

  toJSON(data?: any) {
    data = typeof data === 'object' ? data : {};
    data['id'] = this.id;
    data['productId'] = this.productId;
    data['imageUrl'] = this.imageUrl;
    data['displayOrder'] = this.displayOrder;
    data['isMain'] = this.isMain;
    return data;
  }
}

export interface IProductImageDto {
  id?: string;
  productId?: string;
  imageUrl?: string | undefined;
  displayOrder?: number;
  isMain?: boolean;
}